requirements.txt
fastapi
uvicorn
pandas
psycopg2-binary
python-dotenv
requests
beautifulsoup4
selenium
